remove linux platform specific code for cpu
